Revert "Semi-structured 2:4 sparsity via SparseSemiStructuredTensor"
