+++ vllm/model_executor/layers/sparsity/base_config.py
@@ -2,19 +2,12 @@ from abc import ABC, abstractmethod
 from typing import Any, Dict, List
 
 import torch
-from typing import Type
 
 from vllm.model_executor.layers.linear import LinearMethodBase
-from magic_wand import CompressedStorageFormat
 
 
 class SparsityConfig(ABC):
     """Base class for sparsity configs."""
-
-    @abstractmethod
-    def get_storage_format_cls(self) -> Type[CompressedStorageFormat]:
-        """Sparse representation format"""
-        raise NotImplementedError
 
     @abstractmethod
     def get_name(self) -> str:
