+++ vllm/model_executor/layers/sparsity/sparse_w16a16.py
@@ -1,11 +1,11 @@
-from typing import Any, Dict, List, Type
+from typing import Any, Dict, List, Optional
 
 import torch
+import torch.nn.functional as F
 
+from vllm.model_executor.layers.linear import LinearMethodBase, set_weight_attrs
 from vllm.model_executor.layers.sparsity.base_config import SparsityConfig
-
-from .sparse_w16a16_linear_method import SparseW16A16LinearMethod
-from magic_wand import (CompressedStorageFormat, SparseBitmaskStorageFormat)
+from vllm.model_executor.layers.parameters import SparseParameter
 
 
 class SparseW16A16Config(SparsityConfig):
@@ -20,10 +20,6 @@ class SparseW16A16Config(SparsityConfig):
 
     def __repr__(self) -> str:
         return "SparseW16A16Config()"
-
-    @classmethod
-    def get_storage_format_cls(cls) -> Type[CompressedStorageFormat]:
-        return SparseBitmaskStorageFormat
 
     @classmethod
     def get_name(cls) -> str:
@@ -47,4 +43,57 @@ class SparseW16A16Config(SparsityConfig):
         return cls()
 
     def get_linear_method(self) -> "SparseW16A16LinearMethod":
-        return SparseW16A16LinearMethod(self, self.get_storage_format_cls())
+        return SparseW16A16LinearMethod(self)
+
+
+class SparseW16A16LinearMethod(LinearMethodBase):
+    """Linear method for Sparse W16A16.
+
+    Args:
+        sparsity_config: The sparse config.
+    """
+
+    def __init__(self, sparsity_config: SparseW16A16Config):
+        self.sparsity_config = sparsity_config
+
+    def create_weights(
+        self,
+        input_size_per_partition: int,
+        output_size_per_partition: int,
+        input_size: int,
+        output_size: int,
+        params_dtype: torch.dtype,
+    ) -> Dict[str, Any]:
+        weight = SparseParameter(
+            shape=torch.Size(
+                (output_size_per_partition, input_size_per_partition)),
+            dtype=params_dtype,
+        )
+
+        set_weight_attrs(weight, {"input_dim": 1, "output_dim": 0})
+
+        return {"weight": weight}
+
+    def apply_weights(
+        self,
+        weights: Dict[str, Any],
+        x: torch.Tensor,
+        bias: Optional[torch.Tensor] = None,
+    ) -> torch.Tensor:
+        sparse_weight = weights["weight"]
+
+        # Uncompress to dense
+        dense_weight = sparse_weight.to_dense()
+
+        # # Uncomment to verify sparsity
+        # density = torch.count_nonzero(
+        #     dense_weight).item() / dense_weight.numel()
+        # print(f"sparsity = {1.0 - density}")
+
+        # Standard matrix multiply
+        if bias is not None:
+            output = F.linear(x, dense_weight, bias)
+        else:
+            output = F.linear(x, dense_weight)
+
+        return output
